+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # 📋 DJ Cloudy changelog
 Welcome to DJ Cloudy's release notes / change log! As the name of this file says this is a file when we log changes. What you can find here is some informations about latest releases
 
+## Version pre-1.2.0
+Version 1.2.0 brings the configuration update to us, but this is just a pre-release. What is added:
+- Configuration utils file (utils/configuration.py) with handler class
+    - Handler has these functions: load profile, set key to value, update profile with default, reset key to default, reset profile to default
+- Configuration profile are split to 2 diffrent types: *guild* and *user*
+- Added command `/config view` alongside with the `config` command group
+
+More commands coming in further updates. Stay tuned!
+
 ## Version v1.1.2
 This version brings some QoL (Quality of Life) functionality. First of all - playlist renaming system has been added. If you're bored of your old playlist name just rename it. To do so use the `/playlist rename command`. Second change is now the "star" button in the song controls toggles between adding and removing song from starred playlist. For example: you click first time - checking... no, this song is not in the starred playlist, so the bot adds it. Second click - checking... yes, song is in the playlist, bot removes it.